Work around Gatsby's CSS issues in production
The CSS in production doesn't behave the same as in development. The
main problem is that hovering over links appends CSS from that route,
but that CSS is already included and it messes up the specificity.

The only workaround that I know of is avoiding Gatsby's Link component
until that issue gets resolved. Too bad, prefetching was pretty cool.

gatsbyjs/gatsby#7517