Replace react-icons with manual <svg> <defs>
These icons are taken from IcoMoon. This is actually quite nice because
I get to choose exactly the icons I want. Since id is a string and there
isn't a very straightforward way of knowing that we made a typo, I
improvized some kind of a prop validation. I couldn't use propTypes
because validation is interrogating the DOM.

But, most importantly, this brings the JS size down by ~370 kB. I
couldn't make tree-shaking work with react-icons.