Split Jest mocking post into two posts
I added a new meta field for deciding whether to list the post in the
blog and show it in search results. I'm using this when I want to
preserve the URL, but redirect people somewhere else where a HTTP
redirect isn't applicable. In this case it's because people have to
choose where they want to go.