Add support for series
Series of posts have a special API:

  - if multiple MDX files are added to a directory, they are now parts
  of a series
  - `series.yml` should be added to that directory containing metadata
  about the series (currently only the title of the series)

A series will be displayed as a single item in the list of posts, its
excerpt will be taken from the first part, followed by the list of all
parts. Each part will say that it's a series and contain a list of other
parts, marking which part you're reading now. The exported `title` will
be displayed as a subtitle, below the title of the series.