Make dark mode test stable
Locally the test was passing just because I happened to be testing at
night, when my OS had dark mode turned on anyway, but CI was failing
because it didn't. The problem was that after `cy.reload()`
window.matchMedia was no longer stubbed, so stubbing it on
`window:before:load` fixes the issue.